打折优惠信息设置接口轻化—— 添加打折优惠模块异常

+++ src/main/java/com/vtmer/yisanbang/service/impl/DiscountServiceImpl.java
@@ -1,5 +1,7 @@
 package com.vtmer.yisanbang.service.impl;
 
+import com.vtmer.yisanbang.common.exception.service.discount.DiscountExistException;
+import com.vtmer.yisanbang.common.exception.service.discount.DiscountNotFoundException;
 import com.vtmer.yisanbang.domain.Discount;
 import com.vtmer.yisanbang.mapper.DiscountMapper;
 import com.vtmer.yisanbang.service.DiscountService;
@@ -13,19 +15,25 @@ public class DiscountServiceImpl implements DiscountService {
     private DiscountMapper discountMapper;
 
     @Override
-    public int insert(Discount discount) {
+    public void insert(Discount discount) {
         Boolean res = discountMapper.checkExist();
         if (!res) {
-            return discountMapper.insert(discount);
-        } else {  // 如果已经有打折数据了，返回-1
-            return -1;
+            // 还没有打折优惠设置
+            discountMapper.insert(discount);
+        } else {  // 已经有打折数据了
+            throw new DiscountExistException();
         }
-
     }
 
     @Override
-    public int update(Discount discount) {
-        return discountMapper.update(discount);
+    public void update(Discount discount) {
+        Discount checkExist = selectDiscount();
+        if (checkExist == null) {
+            // 如果不存在优惠信息设置
+            throw new DiscountNotFoundException();
+        }
+        // 存在，更新
+        discountMapper.update(discount);
     }
 
     @Override
@@ -35,6 +43,11 @@ public class DiscountServiceImpl implements DiscountService {
 
     @Override
     public int deleteDiscount() {
+        Discount checkExist = selectDiscount();
+        if (checkExist == null) {
+            // 如果不存在优惠信息设置
+            throw new DiscountNotFoundException();
+        }
         return discountMapper.delete();
     }
 }
